pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dynamo-infer"
version = "0.1.0"
description = "A modular framework for simulating, visualizing, and inferring dynamics of interacting systems"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
keywords = ["dynamics", "inference", "simulation", "physics", "geometric-algebra"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Physics",
    "Topic :: Scientific/Engineering :: Mathematics",
]

dependencies = [
    "jax>=0.4.0",
    "jaxlib>=0.4.0",
    "diffrax>=0.7.0",
    "equinox>=0.11.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "matplotlib>=3.6.0",
    "plotly>=5.20.0",
    "scikit-learn>=1.3.0",
    "pandas>=2.0.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "typer>=0.9.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "isort>=5.12",
    "flake8>=6.0",
    "mypy>=1.0",
    "pre-commit>=3.0",
]
visualization = [
    "polyscope>=2.3.0",
    "bpy>=4.0.0",  # Blender Python API
]
notebooks = [
    "jupyter>=1.0",
    "ipywidgets>=8.0",
    "seaborn>=0.12",
]
docs = [
    "sphinx>=6.0",
    "sphinx-rtd-theme>=1.2",
    "myst-parser>=2.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/dynamo-infer"
Documentation = "https://dynamo-infer.readthedocs.io"
Repository = "https://github.com/yourusername/dynamo-infer"
"Bug Tracker" = "https://github.com/yourusername/dynamo-infer/issues"

[project.scripts]
dynamo-infer = "dynamo_infer.cli:main"
dynamo-simulate = "dynamo_infer.cli:simulate"
dynamo-infer-dynamics = "dynamo_infer.cli:infer"
dynamo-evaluate = "dynamo_infer.cli:evaluate"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --cov=dynamo_infer --cov-report=term-missing"